Generate CSV Bugs
During testing of the bookings feature, it was found that the generate csv function was unreliable due to how tables were joined. This was fixed by doing a left outer join on the invigilator table from the booking table. Also, during testing, it was found that the CSV was not showing values as required. The Exam Returned and Exam Received fields were showing 0's and 1's and not Y's and N's. Logic was added to the exam export endpoint to handle these requirements. Thorough testing was completed to ensure that CSV files were returning correct values for one and many exams in different month periods.