Generate CSV Additional Filter Options
During testing of the generate csv modal, it was found that additional filters would be required. John McColl specifically asked for an "All Exams" option and "All Non-ITA Exams" option to be added to the filter options. These filters were added and then tested by adding multiple different exams were added in the span of a month, and then making sure that the csv generated for reported had the exact number of exams reported back that existed in the bookings calendar for that month.