Financial Liaison Role
In order for staff that are non-csr/ga/liaison to access the bookings feature to generate financial reporting csvs, they needed to be given a role to access the generate financial reporting button. The generate csv button now only shows up if a user has the finance_designate field on the csr table set to 1. If the financial designate field is set to 1, the user will not be able to see any of the create exam buttons (ita, non-ita, group, pesticide).