Appointments - Page and Navigation
As required by clients, the bookings feature needs a separate page for an appointments calendar. To begin development on this, a component for the page and its buttons needed to be created, and additions to the application router and nav bar needed to be made in order to navigate to this new page. Future development for this page will include implementing a calendar component on the page that has similar functionality to the exams calendar, but implements black out periods.