Login Notifications
Upon login, if a user existed in keycloak, but not in the application csr table they were supposed to see a warning notification telling them to contact an administrator. However, this alert notification was broken, and not working properly. This PR contains the fix this to this issue, as well as updating the displayed message shown to the user when this condition is met.